gitlabtest/pyproject.toml
[project]
name = "brand-automations"
version = "0.1.0"
description = "Brand Boekhouders - HubSpot/Wefact/Clockify automation API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104",
    "hubspot-api-client>=8.1",
    "pydantic>=2.5",
    "uvicorn>=0.15",
    "python-dotenv>=0.21",
    "sentry-sdk[fastapi]",
    "pandas>=2.2.3",
    "matplotlib>=3.9",
    "python-multipart",
    "aiolimiter",
    "apscheduler",
    "requests",
    "openpyxl>=3.1.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "pre-commit>=4.0",
    "httpx>=0.27",
    "types-requests>=2.32.4.20260107",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

# ==== coverage ====
[tool.coverage.run]
source = ["app"]
omit = ["*/scripts/*"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true      # hubspot-api-client has no stubs
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

# ==== ruff ====
[tool.ruff]
target-version = "py312"
line-length = 88
extend-exclude = ["scripts/*"]

[tool.ruff.lint]
select = [
    "F",      # pyflakes
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "C90",    # mccabe complexity
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "ASYNC",  # async rules
    "S",      # flake8-bandit (security)
    "BLE",    # blind exception
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "C4",     # flake8-comprehensions
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "EXE",    # flake8-executable
    "FA",     # flake8-future-annotations
    "ISC",    # implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "PIE",    # flake8-pie
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "INT",    # flake8-gettext
    "PGH",    # pygrep-hooks
    "PL",     # pylint
    "TRY",    # tryceratops
    "FLY",    # flynt
    "PERF",   # perflint
    "RUF",    # ruff-specific rules
]
ignore = [
    "S101",    # Use of assert detected
    "RUF012",  # Mutable class attributes should be annotated with ClassVar
    "SIM102",  # Sometimes nested ifs are more readable
    "SIM117",  # Use single with statement - nested is more readable sometimes
    # UP038 removed in newer ruff versions
    "COM812",  # Trailing comma missing - handled by formatter
    "TID252",  # Prefer absolute imports - allow relative imports
    "G004",    # Logging statement uses f-string - f-strings are more readable
    "TRY300",  # Consider moving statement to else - sometimes clearer as-is
    "TRY401",  # Redundant exception in logging.exception - explicit is better
    "PLR0913", # Too many arguments - sometimes necessary
    "BLE001",  # Blind exception catch - acceptable in top-level handlers
    "PERF401", # List comprehension - loop is clearer for complex operations
    "E501",    # Line too long - let formatter handle it
    "UP007",   # Use X | Y for union types - keep Optional for clarity
    "B008",    # Do not perform function call in default arg (FastAPI Depends pattern)
    "ISC001",  # Implicit string concat - conflicts with formatter
    "PLR2004", # Magic value comparisons - too noisy for existing codebase
    "S113",    # Requests call without timeout - existing API wrapper pattern
    "N806",    # Variable in function should be lowercase - used for class instances
    "C901",    # Too complex - existing functions, refactoring out of scope
    "TRY003",  # Long messages in exceptions - readable error messages preferred
    "PLR0912", # Too many branches - existing functions, refactoring out of scope
    "PLR0911", # Too many return statements - existing functions
    "PLR0915", # Too many statements - existing functions
    "PLW0603", # Global statement usage - intentional module-level state
    "TRY301",  # Abstract raise to inner function - overly opinionated
    "RUF003",  # Ambiguous character in comment - Dutch text with special chars
    "PLC0415", # Import not at top-level - intentional lazy/conditional imports
    "N802",    # Function name not lowercase - existing naming convention
    "S311",    # Pseudo-random generators - not security-critical context
    "E402",    # Module import not at top - scripts with sys.path manipulation
    "RUF006",  # Store asyncio task reference - existing async patterns
    "ASYNC251", # asyncio.sleep(0) - existing patterns
    "TRY203",  # Remove exception handler - clearer explicit re-raise
    "A001",    # Shadowing builtin - existing function signatures
    "A002",    # Shadowing builtin argument - existing function signatures
    "S110",    # try-except-pass - intentional suppression in specific contexts
    "RUF005",  # Use list concatenation - existing patterns
]

[tool.ruff.lint.per-file-ignores]
# Test files can be more permissive
"tests/*.py" = [
    "PLR2004", # Magic values (OK in test data)
    "S105",    # Possible hardcoded password
    "S106",    # Possible hardcoded password in argument
]
"**/test_*.py" = [
    "PLR2004", # Magic values (OK in test data)
    "S105",    # Possible hardcoded password
    "S106",    # Possible hardcoded password in argument
]

[tool.ruff.lint.isort]
force-single-line = true
